backend/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
attrs==23.1.0
audioread==3.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
decorator==5.1.1
fastapi==0.92.0
frozenlist==1.4.1
google-api-core==2.15.0
google-auth==2.25.2
google-cloud-aiplatform==1.38.1
google-cloud-bigquery==3.14.1
google-cloud-core==2.4.1
google-cloud-resource-manager==1.11.0
google-cloud-speech==2.23.0
google-cloud-storage==2.14.0
google-cloud-texttospeech==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httptools==0.6.1
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.351
langchain-community==0.0.4
langchain-core==0.1.1
langsmith==0.0.72
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
marshmallow==3.20.1
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
numba==0.58.1
numpy==1.26.2
packaging==23.2
platformdirs==4.1.0
pooch==1.8.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.28.2
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.23
starlette==0.25.0
tenacity==8.2.3
threadpoolctl==3.2.0
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
vertexai==0.0.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
libsndfile1